+++ weaver/vault/__init__.py
@@ -18,8 +18,4 @@ def includeme(config):
     settings = get_settings(config)
     if asbool(settings.get("weaver.vault", True)):
         LOGGER.info("Adding file vault...")
-        config.add_route(**sd.service_api_route_info(sd.vault_service, settings))
-        config.add_route(**sd.service_api_route_info(sd.vault_file_service, settings))
-        config.add_view(v.upload_file, route_name=sd.vault_service.name, request_method="POST")
-        config.add_view(v.describe_file, route_name=sd.vault_file_service.name, request_method="HEAD")
-        config.add_view(v.download_file, route_name=sd.vault_file_service.name, request_method="GET")
+        config.include("weaver.vault.views")